+++ config.js
@@ -53,7 +53,7 @@ const config = {
     __TEST__,
     __PROD__,
   },
-  compiler_hash_type: __PROD__ ? 'chunkhash' : 'hash',
+  compiler_hash_type: 'hash',
   compiler_fail_on_warning: __TEST__ || __PROD__,
   compiler_output_path: paths.base(envConfig.dir_docs_dist),
   compiler_public_path: '/',
@@ -75,19 +75,6 @@ const config = {
     chunksSort: '', // (string) sort the chunks by that field
     assetsSort: '', // (string) sort the assets by that field
   },
-  compiler_vendor: [
-    '@babel/standalone',
-    'brace',
-    'brace/mode/jsx',
-    'brace/mode/html',
-    'brace/theme/tomorrow_night',
-    'classnames',
-    'copy-to-clipboard',
-    'faker',
-    'react',
-    'react-ace',
-    'react-dom',
-  ],
 }
 
 module.exports = config
